WIIS-2767: New dispatch fields and rights

+++ src/Controller/DispatchController.php
@@ -17,6 +17,7 @@ use App\Entity\Pack;
 use App\Entity\DispatchPack;
 use App\Entity\PieceJointe;
 use App\Entity\Statut;
+use App\Entity\Transporteur;
 use App\Entity\Type;
 use App\Entity\Utilisateur;
 
@@ -49,8 +50,7 @@ use Twig\Error\SyntaxError;
 /**
  * @Route("/acheminements")
  */
-Class DispatchController extends AbstractController
-{
+class DispatchController extends AbstractController {
     /**
      * @var UserService
      */
@@ -70,8 +70,7 @@ Class DispatchController extends AbstractController
      * @param EntityManagerInterface $entityManager
      * @return RedirectResponse|Response
      */
-    public function index(EntityManagerInterface $entityManager)
-    {
+    public function index(EntityManagerInterface $entityManager) {
         if (!$this->userService->hasRightFunction(Menu::DEM, Action::DISPLAY_ACHE)) {
             return $this->redirectToRoute('access_denied');
         }
@@ -85,9 +84,9 @@ Class DispatchController extends AbstractController
         $fieldsParam = $fieldsParamRepository->getByEntity(FieldsParam::ENTITY_CODE_DISPATCH);
 
         return $this->render('dispatch/index.html.twig', [
-			'statuts' => $statutRepository->findByCategorieName(CategorieStatut::DISPATCH, true),
+            'statuts' => $statutRepository->findByCategorieName(CategorieStatut::DISPATCH, true),
             'types' => $types,
-			'modalNewConfig' => [
+            'modalNewConfig' => [
                 'fieldsParam' => $fieldsParam,
                 'dispatchDefaultStatus' => $statutRepository->getIdDefaultsByCategoryName(CategorieStatut::DISPATCH),
                 'typeChampsLibres' => array_map(function (Type $type) use ($champLibreRepository) {
@@ -98,7 +97,7 @@ Class DispatchController extends AbstractController
                         'champsLibres' => $champsLibres,
                     ];
                 }, $types),
-			    'notTreatedStatus' => $statutRepository->findByCategorieName(CategorieStatut::DISPATCH, true, true),
+                'notTreatedStatus' => $statutRepository->findByCategorieName(CategorieStatut::DISPATCH, true, true),
             ]
         ]);
     }
@@ -113,8 +112,7 @@ Class DispatchController extends AbstractController
      * @throws SyntaxError
      */
     public function api(Request $request,
-                        DispatchService $dispatchService): Response
-    {
+                        DispatchService $dispatchService): Response {
         if ($request->isXmlHttpRequest()) {
 
             if (!$this->userService->hasRightFunction(Menu::DEM, Action::DISPLAY_ACHE)) {
@@ -144,14 +142,19 @@ Class DispatchController extends AbstractController
                         DispatchService $dispatchService,
                         AttachmentService $attachmentService,
                         EntityManagerInterface $entityManager,
-                        TranslatorInterface $translator): Response
-    {
+                        TranslatorInterface $translator): Response {
         if ($request->isXmlHttpRequest()) {
+            if (!$this->userService->hasRightFunction(Menu::DEM, Action::CREATE) ||
+                !$this->userService->hasRightFunction(Menu::DEM, Action::CREATE_ACHE)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
             $post = $request->request;
             $statutRepository = $entityManager->getRepository(Statut::class);
             $typeRepository = $entityManager->getRepository(Type::class);
             $emplacementRepository = $entityManager->getRepository(Emplacement::class);
             $utilisateurRepository = $entityManager->getRepository(Utilisateur::class);
+            $transporterRepository = $entityManager->getRepository(Transporteur::class);
 
             $dispatch = new Dispatch();
             $date = new DateTime('now', new \DateTimeZone('Europe/Paris'));
@@ -163,8 +166,12 @@ Class DispatchController extends AbstractController
             $comment = $post->get('commentaire');
             $startDateRaw = $post->get('startDate');
             $endDateRaw = $post->get('endDate');
+            $transporter = $post->get('transporter');
+            $transporterTrackingNumber = $post->get('transporterTrackingNumber');
+            $commandNumber = $post->get('commandNumber');
             $receiver = $post->get('receiver');
             $emergency = $post->get('emergency');
+            $project = $post->get('project');
 
             $startDate = !empty($startDateRaw) ? $dispatchService->createDateFromStr($startDateRaw) : null;
             $endDate = !empty($endDateRaw) ? $dispatchService->createDateFromStr($endDateRaw) : null;
@@ -198,12 +205,28 @@ Class DispatchController extends AbstractController
                 $dispatch->setEndDate($endDate);
             }
 
+            if (!empty($transporter)) {
+                $dispatch->setTransporter($transporterRepository->find($transporter) ?? null);
+            }
+
+            if (!empty($transporterTrackingNumber)) {
+                $dispatch->setTransporterTrackingNumber($transporterTrackingNumber);
+            }
+
+            if (!empty($commandNumber)) {
+                $dispatch->setCommandNumber($commandNumber);
+            }
+
             if (!empty($receiver)) {
                 $dispatch->setReceiver($utilisateurRepository->find($receiver) ?? null);
             }
 
             if (!empty($emergency)) {
                 $dispatch->setUrgent($post->getBoolean('urgent'));
+            }
+
+            if (!empty($project)) {
+                $dispatch->setCommandNumber($project);
             }
 
             $freeFieldService->manageFreeFields($dispatch, $post->all(), $entityManager);
@@ -233,7 +256,7 @@ Class DispatchController extends AbstractController
             return new JsonResponse([
                 'success' => true,
                 'redirect' => $this->generateUrl('dispatch_show', ['id' => $dispatch->getId()]),
-                'msg' => $translator->trans("acheminement.L''acheminement a bien été créé") . '.'
+                'msg' => $translator->trans('acheminement.L\'acheminement a bien été créé') . '.'
             ]);
         }
         throw new NotFoundHttpException('404 not found');
@@ -248,8 +271,7 @@ Class DispatchController extends AbstractController
      */
     public function show(Dispatch $dispatch,
                          EntityManagerInterface $entityManager,
-                         DispatchService $dispatchService)
-    {
+                         DispatchService $dispatchService, TranslatorInterface $translator) {
         if (!$this->userService->hasRightFunction(Menu::DEM, Action::DISPLAY_ACHE)) {
             return $this->redirectToRoute('access_denied');
         }
@@ -286,14 +308,13 @@ Class DispatchController extends AbstractController
      */
     public function printDispatchStateSheet(Dispatch $dispatch,
                                             PDFGeneratorService $PDFGenerator,
-                                            TranslatorInterface $translator): ?Response
-    {
+                                            TranslatorInterface $translator): ?Response {
         if ($dispatch->getDispatchPacks()->isEmpty()) {
             throw new NotFoundHttpException($translator->trans("acheminement.La fiche d\'état n\'existe pas pour cet acheminement"));
         }
 
         $packsConfig = $dispatch->getDispatchPacks()
-            ->map(function(DispatchPack $dispatchPack) use ($dispatch, $translator){
+            ->map(function (DispatchPack $dispatchPack) use ($dispatch, $translator) {
                 return [
                     'title' => 'Acheminement n°' . $dispatch->getId(),
                     'code' => $dispatchPack->getPack()->getCode(),
@@ -302,12 +323,12 @@ Class DispatchController extends AbstractController
                         'Date de validation' => $dispatch->getValidationDate() ? $dispatch->getValidationDate()->format('d/m/Y H:i:s') : '',
                         'Demandeur' => $dispatch->getRequester() ? $dispatch->getRequester()->getUsername() : '',
                         'Destinataire' => $dispatch->getReceiver() ? $dispatch->getReceiver()->getUsername() : '',
-                        $translator->trans('acheminement.emplacement dépose') => $dispatch->getLocationTo() ? $dispatch->getLocationTo()->getLabel() : '',
-                        $translator->trans('acheminement.emplacement prise') => $dispatch->getLocationFrom() ? $dispatch->getLocationFrom()->getLabel() : ''
+                        $translator->trans('acheminement.Emplacement dépose') => $dispatch->getLocationTo() ? $dispatch->getLocationTo()->getLabel() : '',
+                        $translator->trans('acheminement.Emplacement prise') => $dispatch->getLocationFrom() ? $dispatch->getLocationFrom()->getLabel() : ''
                     ]
                 ];
             })
-        ->toArray();
+            ->toArray();
 
         $fileName = 'Etat_acheminement_' . $dispatch->getId() . '.pdf';
         return new PdfResponse(
@@ -330,14 +351,19 @@ Class DispatchController extends AbstractController
                          FreeFieldService $freeFieldService,
                          EntityManagerInterface $entityManager,
                          TranslatorInterface $translator): Response {
-
         $statutRepository = $entityManager->getRepository(Statut::class);
         $dispatchRepository = $entityManager->getRepository(Dispatch::class);
         $utilisateurRepository = $entityManager->getRepository(Utilisateur::class);
+        $transporterRepository = $entityManager->getRepository(Transporteur::class);
         $emplacementRepository = $entityManager->getRepository(Emplacement::class);
 
         $post = $request->request;
         $dispatch = $dispatchRepository->find($post->get('id'));
+
+        if (!$this->userService->hasRightFunction(Menu::DEM, Action::EDIT) ||
+            !$dispatch->getStatut()->isTreated() && !$this->userService->hasRightFunction(Menu::DEM, Action::EDIT_UNPROCESSED_DISPATCH)) {
+            return $this->redirectToRoute('access_denied');
+        }
 
         $startDateRaw = $post->get('startDate');
         $endDateRaw = $post->get('endDate');
@@ -351,8 +377,7 @@ Class DispatchController extends AbstractController
         if (!$oldStatus || !$oldStatus->getTreated()) {
             $newStatus = $statutRepository->find($post->get('statut'));
             $dispatch->setStatut($newStatus);
-        }
-        else {
+        } else {
             $newStatus = null;
         }
 
@@ -365,17 +390,27 @@ Class DispatchController extends AbstractController
 
         $receiverData = $post->get('receiver');
         $requesterData = $post->get('requester');
+        $transporterData = $post->get('requester');
         $receiver = $receiverData ? $utilisateurRepository->find($receiverData) : null;
         $requester = $requesterData ? $utilisateurRepository->find($requesterData) : null;
+        $transporter = $transporterData ? $transporterRepository->find($transporterData) : null;
+
+        $transporterTrackingNumber = $post->get('transporterTrackingNumber');
+        $commandNumber = $post->get('commandNumber');
+        $project = $post->get('project');
 
         $dispatch
             ->setStartDate($startDate)
             ->setEndDate($endDate)
+            ->setTransporter($transporter)
+            ->setTransporterTrackingNumber($transporterTrackingNumber)
+            ->setCommandNumber($commandNumber)
             ->setRequester($requester)
             ->setReceiver($receiver)
             ->setUrgent($post->getBoolean('urgent'))
             ->setLocationFrom($locationTake)
             ->setLocationTo($locationDrop)
+            ->setProject($project)
             ->setCommentaire($post->get('commentaire') ?: '');
 
         $freeFieldService->manageFreeFields($dispatch, $post->all(), $entityManager);
@@ -412,7 +447,7 @@ Class DispatchController extends AbstractController
                 'showDetails' => $dispatchService->createHeaderDetailsConfig($dispatch)
             ]),
             'success' => true,
-            'msg' => $translator->trans("acheminement.L''acheminement a bien été modifié") . '.'
+            'msg' => $translator->trans('acheminement.L\'acheminement a bien été modifié') . '.'
         ]);
     }
 
@@ -423,8 +458,7 @@ Class DispatchController extends AbstractController
      * @return Response
      */
     public function editApi(Request $request,
-                            EntityManagerInterface $entityManager): Response
-    {
+                            EntityManagerInterface $entityManager): Response {
         if ($request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
             $statutRepository = $entityManager->getRepository(Statut::class);
             $dispatchRepository = $entityManager->getRepository(Dispatch::class);
@@ -435,6 +469,12 @@ Class DispatchController extends AbstractController
             $fieldsParam = $fieldsParamRepository->getByEntity(FieldsParam::ENTITY_CODE_DISPATCH);
 
             $dispatch = $dispatchRepository->find($data['id']);
+
+            if (!$this->userService->hasRightFunction(Menu::DEM, Action::EDIT) ||
+                !$dispatch->getStatut()->isTreated() && !$this->userService->hasRightFunction(Menu::DEM, Action::EDIT_UNPROCESSED_DISPATCH)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
             $json = $this->renderView('dispatch/modalEditContentDispatch.html.twig', [
                 'dispatch' => $dispatch,
                 'fieldsParam' => $fieldsParam,
@@ -457,15 +497,20 @@ Class DispatchController extends AbstractController
      */
     public function delete(Request $request,
                            EntityManagerInterface $entityManager,
-                           TranslatorInterface $translator): Response
-    {
+                           TranslatorInterface $translator): Response {
         if ($request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
             $dispatchRepository = $entityManager->getRepository(Dispatch::class);
             $attachmentRepository = $entityManager->getRepository(PieceJointe::class);
 
             $dispatch = $dispatchRepository->find($data['dispatch']);
 
-            if($dispatch) {
+            if (!$this->userService->hasRightFunction(Menu::DEM, Action::DELETE) ||
+                !$this->userService->hasRightFunction(Menu::DEM, Action::DELETE_ACHE) ||
+                !$dispatch->getStatut()->isTreated() && !$this->userService->hasRightFunction(Menu::DEM, Action::DELETE_UNPROCESSED_DISPATCH)) {
+                return $this->redirectToRoute('access_denied');
+            }
+
+            if ($dispatch) {
                 $attachments = $attachmentRepository->findBy(['dispatch' => $dispatch]);
                 foreach ($attachments as $attachment) {
                     $entityManager->remove($attachment);
@@ -495,8 +540,7 @@ Class DispatchController extends AbstractController
      * @param Request $request
      * @param EntityManagerInterface $entityManager
      */
-    private function persistAttachments(Dispatch $entity, AttachmentService $attachmentService, Request $request, EntityManagerInterface $entityManager)
-    {
+    private function persistAttachments(Dispatch $entity, AttachmentService $attachmentService, Request $request, EntityManagerInterface $entityManager) {
         $attachments = $attachmentService->createAttachements($request->files);
         foreach ($attachments as $attachment) {
             $entityManager->persist($attachment);
@@ -511,8 +555,7 @@ Class DispatchController extends AbstractController
      * @param Dispatch $dispatch
      * @return Response
      */
-    public function apiPack(Dispatch $dispatch): Response
-    {
+    public function apiPack(Dispatch $dispatch): Response {
         return new JsonResponse([
             'data' => $dispatch->getDispatchPacks()
                 ->map(function (DispatchPack $dispatchPack) {
@@ -567,8 +610,7 @@ Class DispatchController extends AbstractController
         if ($alreadyCreated) {
             $success = false;
             $message = $translator->trans('acheminement.Le colis existe déjà dans cet acheminement');
-        }
-        else {
+        } else {
             $natureRepository = $entityManager->getRepository(Nature::class);
             $packRepository = $entityManager->getRepository(Pack::class);
 
@@ -659,8 +701,7 @@ Class DispatchController extends AbstractController
      */
     public function deletePack(Request $request,
                                TranslatorInterface $translator,
-                               EntityManagerInterface $entityManager): Response
-    {
+                               EntityManagerInterface $entityManager): Response {
         if ($request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
             $dispatchPackRepository = $entityManager->getRepository(DispatchPack::class);
 
@@ -693,11 +734,10 @@ Class DispatchController extends AbstractController
                                             EntityManagerInterface $entityManager,
                                             DispatchService $dispatchService,
                                             TranslatorInterface $translator,
-                                            Dispatch $dispatch): Response
-    {
+                                            Dispatch $dispatch): Response {
         $status = $dispatch->getStatut();
 
-        if(!$status || !$status->getTreated()) {
+        if (!$status || !$status->getTreated()) {
             $data = json_decode($request->getContent(), true);
             $statusRepository = $entityManager->getRepository(Statut::class);
 
@@ -712,8 +752,7 @@ Class DispatchController extends AbstractController
                 $loggedUser = $this->getUser();
 
                 $dispatchService->validateDispatchRequest($entityManager, $dispatch, $treatedStatus, $loggedUser);
-            }
-            else {
+            } else {
                 return new JsonResponse([
                     'success' => false,
                     'msg' => "Le statut sélectionné doit être de type traité et correspondre au type de la demande."
@@ -723,7 +762,7 @@ Class DispatchController extends AbstractController
 
         return new JsonResponse([
             'success' => true,
-            'msg' => 'La ' . $translator->trans('acheminement.acheminement') . 'a bien été traité(e).',
+            'msg' => $translator->trans('acheminement.L\'acheminement a bien été traité'),
             'redirect' => $this->generateUrl('dispatch_show', ['id' => $dispatch->getId()])
         ]);
     }
